Finished and Organized PID for realsies.
Remembereed to press the save button before I commit.
Changes the ImageProccesingConstants to ImageProccesingSuppliers, as a more fit name, and put the constants in SubSystemConstants.
Seperated the suppliers into two static classes nested in the ImageProccesingSuppliers - one for the small reflective and one for the big one.
Also wrote the DriveArcadeWithPID command, and made it start at the press of the BACK  button in the buttonJoystick.