Organized Constants
Deleted the command group GetPIDArcadeDrive and instead put it as a DriveArcadeWithPID variable in OI.
Cleaned constants and changed some of the variable names. (added Ks, changed hight to height, and some other things), as well as the ones in the ConstantHandler.
Moved all the PID constants from ImageProccesingSuppliers to a static PID interface in SubsystemConstants.
Changed constants that were double to Supplier<Double>.
Also added a more organized list of all the constants that need to be found.