Added button switches (Untested), cycleCamera
Changed the code such that each button has a switch on th shuffleboard, and the button will only do things when its switch is on.

For this purpose, I created four new classes:

The First, ConstantHandlerEG, inherits from ConstantHandler and adds methods for constant booleans, which did not exist in the original Spikeslib class.

Second, ButtonSwitches is a simple class that contains suppliers for the buttons switches utilizing ConstantHandlerEG, and divided into the subsystems of each button through interfaces.

Third and Fourth, MoveSubsystemWithSwitch  and DriveTankWithSwitch inherit from MoveBasicSubsystem and DriveTank, respectively. Each requires the additonal Supplier<Boolean> for the switches - they only move the subsystem if
the supplier returns true.

Finally, I changed the apropriate commands in OI.

Note that this has not been tested yet, and so is probable to malfunction.

In addition to all of this, I have also deleted SwitchToCameraA and SwitchToCameraB, as they are inefficent and unused, and created a much simpler CycleCameras, which moves the camera in the next port using a camerasNumber constant for the amount of cameras and a currentCamera variable for the current port.